+++ test/render-vector-test.js
@@ -192,9 +192,9 @@ describe('render vector tiles', function () {
       assert.strictEqual(expectedBuffer.length, buffer.length)
       assert.deepStrictEqual(expectedBuffer, buffer)
 
-      assert.strictEqual(1, renderer._mapPool.size)
+      assert.strictEqual(1, renderer.getStats().get('pool.count'))
       await renderer.close()
-      assert.strictEqual(0, renderer._mapPool.size)
+      assert.strictEqual(0, renderer.getStats().get('pool.count'))
     })
   })
 })